stash commit
